+++ spec/javascripts/govuk_publishing_components/analytics-ga4/ga4-core.spec.js
@@ -285,6 +285,14 @@ describe('GA4 core', function () {
       expect(domains).toContain('gov.uk')
     })
 
+    it('standardises search terms for consistency across trackers', function () {
+      var searchTerm = 'NO UPPERCASE, NO %2B plus + signs, NO PII email@example.com SW1A 2AA 1st Jan 1990 and    NO    extra    spaces \n \r      '
+      var expected = 'no uppercase, no plus signs, no pii [email] [postcode] [date] and no extra spaces'
+
+      searchTerm = GOVUK.analyticsGa4.core.trackFunctions.standardiseSearchTerm(searchTerm)
+      expect(searchTerm).toEqual(expected)
+    })
+
     describe('when the data-ga4-set-indexes attribute exists on the module', function () {
       var module